% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiStats.R
\name{make_race_combo}
\alias{make_race_combo}
\title{for a race \code{n} the combo for \verb{race1 == race2 ==n} the combo is is \verb{2n - 1}
for \code{race1 == n && race2 != n} the combo is \verb{2n}}
\usage{
make_race_combo(race1, race2)
}
\arguments{
\item{race1}{a vector of race index for the index}

\item{race2}{a vector of race index for the partners}
}
\value{
a vector of the same size with the race combos
}
\description{
for a race \code{n} the combo for \verb{race1 == race2 ==n} the combo is is \verb{2n - 1}
for \code{race1 == n && race2 != n} the combo is \verb{2n}
}
