% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetStats.R
\name{reweight_age_pyr}
\alias{reweight_age_pyr}
\title{Re-weight age pyramid}
\usage{
reweight_age_pyr(age.pyramid, young.prop, age.break)
}
\arguments{
\item{age.pyramid}{A vector of proportions.}

\item{young.prop}{The weight that should be applied to the portion of
\code{age.pyramid} that is below \code{age.break}}

\item{age.break}{The index at which the upper group should begin.}
}
\description{
This function re-weights a given age pyramid (i.e., a vector
with one element per year in an age range that specifies what proportion of
the population should be  that age) so that the desired fraction of the
population falls below a specified age boundary.
}
\examples{
unif.age.pyr <- rep(1/100, 100)
reweighted.age.pyr <- reweight_age_pyr(unif.age.pyr, 0.995, 65)

}
