% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiStats.R
\name{trim_epistats}
\alias{trim_epistats}
\title{Reduces the Size of the Epistats Object by Trimming its Models}
\usage{
trim_epistats(epistats)
}
\arguments{
\item{epistats}{the \code{epistats} object to be trimmed}
}
\value{
a trimmed \code{epistats}
}
\description{
The \code{epistats} object contains 3 GLM models. These R object take up a lot of
space and only a small subset of their functionalities is used by
EpiModelHIV. This function trims these models to save up space without
compromising EpiModelHIV functionalities.
}
