% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetStats.R
\name{trim_netstats}
\alias{trim_netstats}
\title{Reduces the Size of the Netstats Object by Removing the Networks it Contains}
\usage{
trim_netstats(netstats)
}
\arguments{
\item{netstats}{the \code{netstats} object to be trimmed}
}
\value{
a trimmed \code{netstats}
}
\description{
Once \code{netest} has been built, the network object stored in \code{netstats} are
never used again and take up a lot of space. This function removes them to
reduce the stored size of \code{netstats}. Roughly 10MiB are saved for 100k nodes
networks.
}
