% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARTnet-package.R
\docType{package}
\name{ARTnet-package}
\alias{ARTnet-package}
\alias{ARTnet}
\title{Epidemic and Network Model Parameterization with the ARTnet Dataset}
\description{
\tabular{ll}{
   Package: \tab ARTnet\cr
   Type: \tab Package\cr
   Version: \tab 1.0.0\cr
   Date: \tab 2019-05-23\cr
   License: \tab GPL-3\cr
   LazyLoad: \tab yes\cr
}
}
\keyword{package}
