% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARTnet-package.R
\docType{package}
\name{ARTnet-package}
\alias{ARTnet-package}
\alias{ARTnet}
\title{Epidemic and Network Model Parameterization with the ARTnet Dataset}
\description{
\tabular{ll}{
   Package: \tab ARTnet\cr
   Type: \tab Package\cr
   Version: \tab 2.0.0\cr
   Date: \tab 2019-11-12\cr
   License: \tab GPL-3\cr
   LazyLoad: \tab yes\cr
}
}
\details{
The ARTnet package provides a suite of functions for the parameterization
of epidemic and network models using data from the ARTnet cross-sectional survery of men
who have sex with men (MSM) in the United States.
}
\keyword{package}
