% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiStats.R
\name{build_epistats}
\alias{build_epistats}
\title{Build EpiStats}
\usage{
build_epistats(city_name = "Atlanta")
}
\arguments{
\item{city_name}{}
}
\description{
Build EpiStats
}
\examples{
test <- build_epistats(city_name = "Atlanta")

}
