% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiStats.R
\name{build_epistats}
\alias{build_epistats}
\title{Statistical Models for Act Rates, Condom Use, and Starting HIV Prevalence}
\usage{
build_epistats(
  geog.lvl = NULL,
  geog.cat = NULL,
  race = TRUE,
  age.limits = c(15, 65),
  age.breaks = c(25, 35, 45, 55),
  age.sexual.cessation = NULL,
  init.hiv.prev = NULL,
  time.unit = 7,
  browser = FALSE
)
}
\arguments{
\item{geog.lvl}{Specifies geographic level for ARTnet statistics.}

\item{geog.cat}{Specifies one or more geographic strata within the level to base ARTnet
statistics on. If the vector is of length 2+, data from the strata will be combined into
one analysis.}

\item{race}{If \code{TRUE}, stratify model estimates by race/ethnic grouping.}

\item{age.limits}{Upper and lower limit of age range to subset ARTnet data by. Minimum of 15 and
maximum of 100 allowed.}

\item{age.breaks}{Ages that define the upper closed boundary of the age categories. Default is
\code{c(25, 35, 45, 55)}, which corresponds to \verb{(0, 25], (25, 35], (35, 45], (45, 55], (55, 65]}
with \code{age.limits = c(15, 65)}.}

\item{age.sexual.cessation}{Age of cessation of sexual activity, while aging process continues
through the upper age limit. Maximum allowed value of 65.}

\item{init.hiv.prev}{Initial HIV prevalence to be used in epidemic model estimated model, with a
numerical vector of size 3 corresponding to starting prevalence in three race/ethnic
groups (Black, Hispanic, and White/Other, respectively). If \code{init.hiv.prev = NULL},
\code{build_epistats} will estimate a logistic regression model to predict starting prevalence
as a function of estimated prevalence in ARTnet as a function of race/ethnicity and age.}

\item{time.unit}{Specifies time unit for time-dependent ARTnet statistics. Default is 7,
corresponding to a weekly time unit. Allowed inputs range from 1 for a daily time unit to
30 for a monthly time unit.}

\item{browser}{If \code{TRUE}, run \code{build_epistats} in interactive browser mode.}
}
\description{
Builds statistical models governing act rates and probability of condom use among
main, casual and one-time sexual partnerships, and the probability of diagnosed HIV infection,
for use in the EpiModelHIV workflow.
}
\details{
The \code{build_epistats} function provides a way to input of geographic, age, and racial parameters
necessary to build statistical models predicting sexual activity and condom use in sexual
partnerships among men who have sex with men (MSM). Estimation of these models is performed using
data from the \code{ARTnetData} package, containing the results of the ARTnet study.
\subsection{Explanation of Parameter Values}{
\itemize{
\item \code{geog.lvl}: level of geographic stratification desired. Acceptable values are \code{"city"},
\code{"county"}, \code{"state"}, \code{"region"}, and \code{"division"} corresponding to the metropolitan
statistical area, county, state, census region, and census division, respectively. Default
value is \code{NULL}, indicating no geographic stratification.
\item \code{geog.cat}: given a geographic level above, \code{"geog.cat"} is a vector comprising the desired
feature(s) of interest. Acceptable values are based on the chosen geographic level:
\itemize{
\item \code{city}: \code{"Atlanta"}, \code{"Boston"}, \code{"Chicago"}, \code{"Dallas"}, \code{"Denver"}, \code{"Detroit"},
\code{"Houston"}, \code{"Los Angeles"}, \code{"Miami"}, \code{"New York City"}, \code{"Philadelphia"}, \code{"San Diego"},
\code{"San Francisco"}, \code{"Seattle"}, \code{"Washington DC"}.
\item \code{county}: FIPS codes for the county or county equivalents to be included. Selecting a
single county or a set of smaller counties may lead to insufficient data used to estimate
the models; an error may result.
\item \code{state}: Two-letter postal code for each state.
\item \code{division}: \code{"1"} (New England), \code{"2"} (Middle Atlantic), \code{"3"} (East North Central),
\code{"4"} (West North Central), \code{"5"} (South Atlantic), \code{"6"} (East South Central),
\code{"7"} (West South Central), \code{"8"} (Mountain) \code{"9"} (Pacific).
\item \code{region}: \code{"1"} (Northeast), \code{"2"} (Midwest), \code{"3"} (South), \code{"4"} (North)
}
\item \code{race}: whether to introduce modeling by racial stratification. \code{TRUE} or \code{FALSE}.
\item \code{age.limits}: a vector giving the lower and upper limit for the age of interest. Set to
\code{c(15, 65)} by default. Although the ARTnet data include respondents from age 15 to 65, this
may be set to restricted values within that range (for example, \code{c(25, 40)}) for a subsetted
data analysis. The upper boundary may be set up to age 100 (that is, \code{c(15, 100)}), which may
be used in models where sexual activity ceases before mortality.
\item \code{age.breaks}: a vector giving the upper age breaks to categorize data by age. Must be within
the bounds specified by \code{age.limits}. These should be the interior age breaks only (that is,
there is no need to include the age limit boundaries). If an age of sexual cessation is added,
then this age is also added to the age breaks if it is not explicitly specified.
\item \code{age.sexual.cessation}: a numerical value for the age of cessation of sexual activity. This may
be by assumption or given data constraints (ARTnet eligibility were through age 65, so the
maximum value here is 65). This specification is useful for models in which the HIV transmission
process stops at a certain age but aging and other demographic features should continue through
natural mortality.
\item \code{time.unit}: a number between 1 and 30 that specifies time units for ARTnet statistics. Set to
\code{7} by default.
}
}
}
\examples{
# Age and geographic stratification, for the Atlanta metropolitan statistical area
epistats1 <- build_epistats(geog.lvl = "city",
                            geog.cat = "Atlanta",
                            age.limits = c(20, 50),
                            age.breaks = c(24, 34, 44))

# Default age stratification
epistats2 <- build_epistats(geog.lvl = "state", geog.cat = "WA")

# Default age stratification, multiple states
epistats3 <- build_epistats(geog.lvl = "state", geog.cat = c("ME", "NH", "VT"))

# No race stratification
epistats4 <- build_epistats(geog.lvl = "state", geog.cat = "GA", race = FALSE)

# Age and race stratification, for the municipality (not metro) of New York City
# geog.cat values are FIPS codes for the 5 boroughs of NYC
epistats5 <- build_epistats(geog.lvl = "county",
                            geog.cat = c(36005, 36047, 36061, 36081, 36085),
                            age.limits = c(20, 50),
                            age.breaks = c(24, 34, 44))

# Use broader age range (to age 100) but with sexual cessation at age 65
epistats6 <- build_epistats(geog.lvl = "city",
                            geog.cat = "Atlanta",
                            race = TRUE,
                            age.limits = c(15, 100),
                            age.breaks = c(25, 35, 45, 55, 65),
                            age.sexual.cessation = 65)

}
