% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiStats.R
\name{build_epistats}
\alias{build_epistats}
\title{Build EpiStats}
\usage{
build_epistats(geog.lvl = NULL, geog.cat = NULL, race = TRUE,
  age.lim = c(15, 65), age.bks = c(0, 24, 34, 44, 54, 64, 100),
  browser = FALSE)
}
\arguments{
\item{geog.lvl}{Specifies geographic feature for ARTnet statistics..}

\item{geog.cat}{Specifies geographic stratum to base ARTnet statistics on.}

\item{race}{Whether to stratify by racial status. Default is TRUE.}

\item{age.lim}{Upper and lower limit. Age rage to subset ARTnet data by. Default is 0 to 65}

\item{age.bks}{Ages that define age categories. If NULL, age breaks are equally spaced.}

\item{browser}{Run function in interactive browser mode.}
}
\description{
Build EpiStats
}
\examples{

epistats1 <- build_epistats(geog.lvl = "city", geog.cat = "Atlanta", age.lim = c(15, 65),
age.bks = c(0, 24, 34, 44, 54, 64, 100))

epistats2 <- build_epistats(geog.lvl = "state", geog.cat = "WA")

#No racial stratification
epistats5 <- build_epistats(geog.lvl = "state", geog.cat = "GA", race = FALSE)


}
