% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetParams.R
\name{build_netparams}
\alias{build_netparams}
\title{Build NetParams}
\usage{
build_netparams(epistats)
}
\arguments{
\item{epistats}{Output from \code{\link{build_epistats}}.}
}
\description{
Build NetParams
}
\examples{
epistats <- build_epistats(city_name = "Atlanta")
netparams <- build_netparams(epistats = epistats)

}
