% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetParams.R
\name{build_netparams}
\alias{build_netparams}
\title{Build NetParams}
\usage{
build_netparams(epistats, smooth.main.dur.old = FALSE, browser = FALSE)
}
\arguments{
\item{epistats}{Output from \code{\link{build_epistats}}.}

\item{smooth.main.dur.old}{If \code{TRUE}, average main durations for oldest and second oldest age
groups. FALSE by default.}

\item{browser}{Run function in interactive browser mode.}
}
\description{
Build NetParams
}
\examples{
epistats <- build_epistats(geog.lvl = "state", geog.cat = "GA", race = TRUE, age.lim = c(15, 65),
age.bks = c(20, 30, 40, 50, 60))
netparams <- build_netparams(epistats = epistats, smooth.main.dur.old = TRUE)

}
