% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetParams.R
\name{build_netparams}
\alias{build_netparams}
\title{Calculate Individual-Level Network Parameters}
\usage{
build_netparams(epistats, smooth.main.dur = FALSE, browser = FALSE)
}
\arguments{
\item{epistats}{Output from \code{\link{build_epistats}}.}

\item{smooth.main.dur}{If \code{TRUE}, function averages the main sexual partnership durations for
oldest and second oldest age groups.}

\item{browser}{If \code{TRUE}, run \code{build_netparams} in interactive browser mode.}
}
\description{
Builds statistical models predicting mean degree, mixing, and duration of sexual
partnerships, for use in the EpiModelHIV workflow.
}
\details{
\code{build_netparams} is a helper function that constructs the necessary network parameters for use
in building network models with \code{\link{build_netstats}}, building on models estimated using
\code{\link{build_epistats}}.

The parameter \code{smooth.main.dur} is used when partnership duration and mortality compete in the
eldest age group; in such a case mean duration is averaged over the oldest and second oldest age
groups (as specified by \code{age.breaks} in \code{\link{build_epistats}}). Subsequently, this smoothing is only
done if there are three or more age categories specified.
}
\examples{

# Example of standard model
epistats <- build_epistats(geog.lvl = "state", geog.cat = "GA", race = TRUE,
                           age.limits = c(20, 50),
                           age.breaks = c(20, 30, 40))
netparams <- build_netparams(epistats = epistats, smooth.main.dur = TRUE)

# Example of model with sexual cessation age < age limit
epistats2 <- build_epistats(geog.lvl = "city",
                            geog.cat = "Atlanta",
                            race = TRUE,
                            age.limits = c(15, 100),
                            age.breaks = c(25, 35, 45, 55, 65),
                            age.sexual.cessation = 65)
netparams2 <- build_netparams(epistats2, smooth.main.dur = TRUE)

}
