% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetParams.R
\name{build_netparams}
\alias{build_netparams}
\title{Build NetParams}
\usage{
build_netparams(epistats)
}
\arguments{
\item{city_name}{}
}
\description{
Build NetParams
}
\examples{
test.epistats <- build_epistats(city_name = "Atlanta")
test.netparams <- build_netparams(epistats = test.epistats)

}
