% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetStats.R
\name{build_netstats}
\alias{build_netstats}
\title{Build EpiStats}
\usage{
build_netstats(epistats, netparams, expect.mort = 0.0005762142)
}
\arguments{
\item{epistats}{Output from \code{\link{build_epistats}}.}

\item{netparams}{Output from \code{\link{build_netparams}}.}
}
\description{
Build EpiStats
}
\examples{
epistats <- build_epistats(city_name = "Atlanta")
netparams <- build_netparams(epistats = epistats)
netstats <- build_netstats(epistats, netparams)

}
