% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetStats.R
\name{build_netstats}
\alias{build_netstats}
\title{Calculate Network Target Statistics}
\usage{
build_netstats(
  epistats,
  netparams,
  network.size = 10000,
  expect.mort = 1e-04,
  edges.avg = FALSE,
  race.prop = NULL,
  browser = FALSE
)
}
\arguments{
\item{epistats}{Output from \code{\link{build_epistats}}.}

\item{netparams}{Output from \code{\link{build_netparams}}.}

\item{network.size}{Size of the starting network.}

\item{expect.mort}{Expected average mortality level to pass into \code{\link{dissolution_coefs}} function.}

\item{edges.avg}{If \code{TRUE}, calculates the overall edges target statistics as a weighted average
of the statistics for edges by race/ethnicity group; if \code{FALSE}, takes the raw average.}

\item{race.prop}{A numerical vector of length 3, containing the proportion of the population with
each of the three values for the nodal attribute "race" in order: White/Other, Black,
and Hispanic).}

\item{browser}{If \code{TRUE}, run \code{build_netparams} in interactive browser mode.}
}
\description{
Calculates the final target statistics for the network models by applying
individual-level network statistics against the population size and structure, for
use in the EpiModelHIV workflow.
}
\details{
This function takes output from \code{\link{build_epistats}} and \code{\link{build_netparams}} to build the relevant
population-level network statistics that will be used in network estimation using the \link{EpiModel}
package.

The parameter \code{edge.avg} allows a user specify how the network edges statistics as estimated from
ARTnet with \code{\link{build_netparams}} should be calculated. With \code{edges.avg = FALSE}, the edges count
is the overall mean degree (divided by 2) times the \code{network.size}. However, this may create model
fitting issues if sample proportions by race/ethnicity in ARTnet do not match the population
proportions (i.e., the overall average may be inconsistent with the weighted average due to
different race denominator sizes). The \code{edges.avg = TRUE} setting calculates the overall edges as
the sum of mean degrees in each race/ethnicity group (divided by 2) times the size of each group.

The \code{race.prop} argument only needs to be specified if the built-in race/ethnicity distribution
data in \code{\link[ARTnetData:race.dist]{ARTnetData::race.dist}} is not used. This may be the case \code{geog.lvl = "county"} or if
\code{geog.cat} has length >1; otherwise, the values can be obtained automatically. If \code{race.prop} is
not supplied in either of these cases, national US values will be used..
}
\examples{
# Standard model with default age stratification
epistats <- build_epistats(geog.lvl = "city", geog.cat = "Atlanta")
netparams <- build_netparams(epistats, smooth.main.dur = TRUE)
netstats <- build_netstats(epistats, netparams)

# Restricted age stratification
epistats2 <- build_epistats(geog.lvl = "state", geog.cat = "GA",
                            age.limits = c(20, 50),
                            age.breaks = c(20, 30, 40))
netparams2 <- build_netparams(epistats2, smooth.main.dur = TRUE)
netstats2 <- build_netstats(epistats2, netparams2)

# Model with sexual cessation age < age limit
epistats3 <- build_epistats(geog.lvl = "city",
                            geog.cat = "Atlanta",
                            race = TRUE,
                            age.limits = c(15, 100),
                            age.breaks = c(25, 35, 45, 55),
                            age.sexual.cessation = 65)
netparams3 <- build_netparams(epistats3, smooth.main.dur = TRUE)
netstats3 <- build_netstats(epistats3, netparams3)

}
