% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetStats.R
\name{build_netstats}
\alias{build_netstats}
\title{Calculate Network Target Statistics}
\usage{
build_netstats(
  epistats,
  netparams,
  network.size = 10000,
  expect.mort = 1e-04,
  edges.avg = FALSE,
  race.prop = NULL
)
}
\arguments{
\item{epistats}{Output from \code{\link{build_epistats}}.}

\item{netparams}{Output from \code{\link{build_netparams}}.}

\item{network.size}{Size of the starting network.}

\item{expect.mort}{Expected average mortality level to pass into
\code{\link{dissolution_coefs}} function.}

\item{edges.avg}{Whether degree differences exist along race. TRUE
or FALSE; default of FALSE.}

\item{race.prop}{The proportion of the population with each of the three
values for the nodal attribute "race" (White.Other, Black, Hispanic).
This only needs to be supplied if geog.lvl = "county" or if geog.cat
has length >1; otherwise, the values can be obtained automatically
from a look-up table. If it is not supplied in either of these cases,
the function will default to national US values.}
}
\description{
Calculates the final target statistics for the network models by
             applying individual-level network statistics against the population
             size and structure, for use in the EpiModelHIV workflow.
}
\details{
\code{build_netstats} takes output from \code{\link{build_epistats}} and
\code{\link{build_netparams}} to build the relevant network statistics
that will be used in network estimation using package \link{EpiModel}.

The parameter \code{edge.avg} allows a user set the network stated edges
to that estimated in \code{\link{build_netparams}} (divided by 2),
with \code{edges.avg = FALSE}, or, if sample proportions do not match
ARTnet population proportions, set to a weighted racial average
with \code{edges.avg = TRUE}.
}
\examples{
epistats <- build_epistats(geog.lvl = "city", geog.cat = "Atlanta")
netparams <- build_netparams(epistats = epistats, smooth.main.dur = TRUE)
netstats <- build_netstats(epistats, netparams)

}
